requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
drf-nested-routers==0.94.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
